remove numpy atomic types. introduce OOP class structure. :art:

+++ grid/grid_provider.py
@@ -14,23 +14,23 @@ class GridProvider:
     :param seed: random seed to be used when option RANDOM is used
     """
 
-    def __init__(self, dimension: np.int8, seed: np.int8 = None, lower_bound: np.float16 = np.float16(-1.),
-                 upper_bound: np.float16 = np.float16(1.)):
+    def __init__(self, dimension: int, seed: int = None, lower_bound: float = -1.,
+                 upper_bound: float = 1.):
         self.dim = dimension
         self.lower_bound = lower_bound
         self.upper_bound = upper_bound
 
-        if isinstance(seed, np.int8):
+        if isinstance(seed, int):
             self.seed = seed
             self.rng = np.random.default_rng(seed=seed)
         else:
             self.rng = np.random.default_rng(seed=None)
 
-    def set_seed(self, seed: np.int8):
+    def set_seed(self, seed: int):
         self.seed = seed
         self.rng = np.random.default_rng(seed=seed)
 
-    def generate(self, grid_type: GridType, scale: np.int32 = None, remove_duplicates: bool = True) -> Grid:
+    def generate(self, grid_type: GridType, scale: int = None, remove_duplicates: bool = True) -> Grid:
         """
         Generate a grid of given type.
         :param scale:  Number of points (per dimension!) when generating the grid equidistantly or randomly.
@@ -55,17 +55,17 @@ class GridProvider:
             points = self._generate_random_grid(num_points=scale)
             return Grid(self.dim, scale, points, grid_type)
 
-    def _generate_random_grid(self, num_points: np.int32) -> np.ndarray:
+    def _generate_random_grid(self, num_points: int) -> np.ndarray:
         return self.rng.uniform(low=self.lower_bound, high=self.upper_bound, size=(num_points, self.dim))
 
-    def _generate_equidistant_grid(self, num_points: np.int32) -> np.ndarray:
+    def _generate_equidistant_grid(self, num_points: int) -> np.ndarray:
         num_points = np.full(shape=self.dim, fill_value=num_points)
         axes = [np.linspace(self.lower_bound, self.upper_bound, num_points[i]) for i in range(self.dim)]
         mesh = np.meshgrid(*axes, indexing='ij')
         return np.stack(mesh, axis=-1).reshape(-1, self.dim)
 
-    def _full_cheby_grid(self, level: np.int32, remove_duplicates: bool = True) -> np.ndarray:
-        grids = [self._uni_grid(np.int32(k)) for k in range(level + 1)]
+    def _full_cheby_grid(self, level: int, remove_duplicates: bool = True) -> np.ndarray:
+        grids = [self._uni_grid(int(k)) for k in range(level + 1)]
 
         memo = {}
         valid_levels = self._valid_combinations(self.dim, level, memo)
@@ -79,7 +79,7 @@ class GridProvider:
             grid = self._remove_duplicates(grid)
         return self._rescale(grid)
 
-    def _valid_combinations(self, d: np.int32, level: np.int32, memo: dict = None):
+    def _valid_combinations(self, d: int, level: int, memo: dict = None):
         if (d, level) in memo:
             return memo[(d, level)]
         if d == 1:
@@ -87,16 +87,16 @@ class GridProvider:
         else:
             result = []
             for current_level in range(level + 1):
-                for sub_combination in self._valid_combinations(np.int32(d - 1), np.int32(level - current_level), memo):
+                for sub_combination in self._valid_combinations(int(d - 1), int(level - current_level), memo):
                     result.append([current_level] + sub_combination)
         memo[(d, level)] = result
         return result
 
-    def _uni_grid(self, level: np.int32) -> np.ndarray:
+    def _uni_grid(self, level: int) -> np.ndarray:
         return np.zeros(1) if level == 0 else self._cheby_nodes(2 ** level + 1)
 
     @staticmethod
-    def _cheby_nodes(n: np.int8) -> np.ndarray:
+    def _cheby_nodes(n: int) -> np.ndarray:
         arr = np.arange(1, n + 1)
         return (-1) * np.cos(np.pi * (arr - 1) / (n - 1))
